+++ src/krill/modules/bootloader-config.ts
@@ -8,54 +8,11 @@
  */
 
 import Sequence from '../krill-sequence'
-import { exec } from '../../lib/utils'
-import Utils from '../../classes/utils'
 
-/**
- *
- * @param this
- */
 export default async function bootloaderConfig(this: Sequence): Promise<void> {
   if (this.distro.familyId === 'debian') {
-      this.bootloaderConfigDebian()
+    this.bootloaderConfigDebian()
   } else if (this.distro.familyId === 'archlinux') {
     this.bootloaderConfigArch() 
   }
-
-  /*
-  // This is common all
-  let cmd =''
-  try {
-    cmd = `chroot ${this.installTarget} sleep 1 ${this.toNull}`
-    await exec(cmd, this.echo)
-  } catch (error) {
-    console.log(error)
-    await Utils.pressKeyToExit(cmd, true)
-  }
-
-  try {
-    cmd = `chroot ${this.installTarget} grub-install ${this.partitions.installationDevice} ${this.toNull}`
-    await exec(cmd, this.echo)
-  } catch (error) {
-    console.log(error)
-    await Utils.pressKeyToExit(cmd, true)
-  }
-
-  try {
-    cmd = `chroot ${this.installTarget} grub-mkconfig -o /boot/grub/grub.cfg ${this.toNull}`
-    await exec(cmd, this.echo)
-  } catch (error) {
-    console.log(error)
-    await Utils.pressKeyToExit(cmd, true)
-  }
-
-  try {
-    cmd = `chroot ${this.installTarget} sleep 1 ${this.toNull}`
-    await exec(cmd, this.echo)
-  } catch (error) {
-    console.log(error)
-    await Utils.pressKeyToExit(cmd, true)
-  }
-  */
 }
-